handlers/face-detection/requirements.txt
boto3==1.34.2
Pillow==10.1.0